Refactor component model stuff
Moved the eloquent queries out of the component and into their own
methods on the Gallery model instead. The lists() method has also
been simplified as the select and get methods are already called
when running lists().

Finally, we move the setting of the gallery var to the end of the
onRender() method. This way we can use the id in $this->page
instead of having to call $this->property().