Fix component when using copied partial code
If you ctrl-double-click the component tag to copy the default
partial code in the October backend, the gallery won't work.

This commit fixes the issue by loading variables onRun, not just
onRender, and adding the gallery object into the component
so that it can be accessed that way if desired.

Also adds a missing __SELF__ in the default partial.